synced all existing routes
