+++ models/Users.js
@@ -71,7 +71,6 @@ const User = sequelize.define(
     },
   },
   {
-    tableName: 'users',
     freezeTableName: true,
     timestamps: false,
   }
